Revert "Moved response parsing into fixture."
This reverts commit 1d883ebd00b227b5eb934204eeb086fac83859df.
